+++ support-frontend/assets/helpers/paymentIntegrations/oneOffContributions.js
@@ -59,7 +59,6 @@ export type StripeChargeData = {|
   paymentData: {
     currency: IsoCurrency,
     amount: number,
-    token: string, // TODO - remove this field once it has been removed from payment-api
     email: string,
     stripePaymentMethod: StripePaymentMethod,
   },
